Add a notebook that reproduces the results from the ComplEx paper (#901)
This adds a notebook that reproduces the results from the ComplEx paper
(http://jmlr.org/proceedings/papers/v48/trouillon16.pdf), in Table 2 in
particular.

In particular, the notebook does an evaluation of the "raw" ranks of a edge `(s,
r, o)` (source, relation type, object) by comparing the model's predicted score
for that edge against the scores for all "modified-source" versions of that link
`(n, r, o)` (for every node `n`), and for all "modified-object" versions `(s, r,
n)`. It does this on the WN18 and FB15k datasets.

The notebook is phrased as a bit of a tutorial for WN18, but omits the
description for FB15k, since the code is otherwise identical. I decided against
defining a function to do all the work for pedagogical reasons: it's easier to
interleave the discussion of the code (and the results) if it's not all in a
single big function that is used for both WN18 and FB15k. (This somewhat relates
to #910; this notebook is both a how-to and a paper reproduction.)

**Why only raw ranks?** The paper (like most knowledge graph papers) focuses
more on the "filtered" ranks, which are ranking a link `(s, r, o)` against only
the modified-source and modified-object instances that aren't known edges (that
is, that don't appear in the train, test or validation set). However, as yet,
I've been unable to reproduce the large jump in performance one sees when
comparing only these edges, despite trying (and manually validating) several
methods of ignoring the known edges. (Follow-up issue: #1060)

**Why so much custom ranking-computing code?** The naive approach to implement
this validation would be to create DataFrames storing all of the modified-source
and modified-object links and running `model.predict` to get a score for each of
the modified edges. On WN18, this would mean:

    2 (modifying source and object, separately) × 5000 (edges in the test set)
        = 10000 `predict` calls

Each of these is on 40943 triples (total number of nodes, to slot in to the
modified source or object column). This is inefficient, and it's much better to
work with the underlying embedding matrices directly, to phrase computing scores
against all nodes as bulk operations on simple matrices (rather than having to
go through embedding layers, etc.). This is implemented in the
`ComplEx.rank_edges_against_all_nodes` function, and the optimised form is
validated to match the naive/manual form in the unit tests.

**Why work with WN18 and FB15k, since those datasets have test-set leakage of
inverse relations?** It's what the paper uses, and so for validating our results
against the authors, we should match.

See: #862