Optimise Cluster-GCN adjacency normalization (#1095)
This cuts the time for training in the demo notebook from ~60ms/step to
~30ms/step, and reduces the time for the included benchmark of the time to
iterate over a whole epoch (times in milliseconds):

|  q |    code |  min | median |  max |
|---:|--------:|-----:|-------:|-----:|
|  1 | this PR | 18.3 |   19.2 | 23.5 |
|    | develop | 38.7 |   40.6 | 50.8 |
|  2 | this PR |  9.7 |   10.4 | 14.9 |
|    | develop | 26.9 |   28.2 | 35.3 |
| 10 | this PR |  3.4 |    3.8 |  6.5 |
|    | develop | 19.7 |   21.2 | 30.0 |

Before this patch, normalizing the adjacency matrix was a huge chunk of the time
required. In particular the `setdiag` call (which resulted in some
`SparseEfficiencyWarning`s) was about 60% of the time of each `__getitem__`
call, based on profiling, and the time in `StellarGraph.to_adjacency_matrix` was
only 26% of the `__getitem__` time.

After this patch, the total time for normalisation decreases to 13% of the new
overall time (that is, even accounting for that overall time being shorter), the
time in `to_adjacency_matrix` has increased to 72%, and the total time for
normalisation has decreased to 13%. All in all, since nothing else chanced, this
means the time to normalise has gotten about 10&times; faster. (This matches
independent benchmarks I did while experimenting with the best way to rewrite
this.)

I investigated other places to put the densification (the `.toarray()` call),
such as after computing `diag` (not noticeably faster) or before doing the row
multiplication (much slower as the size of each cluster increases, with constant
density).

This doesn't try to share code with `core.utils.normalize_adj` (yet) because
that is designed to output sparse matrices, doesn't handle self loops and
doesn't handle the lambda diagonal enhancement, and thus there may not actually
be much code to share.

See: #1094