Compute filtered ranks for evaluating ComplEx (#1080)
This adds the final piece of the evaluation required to make `ComplEx`
non-`@experimental`. it extends the ranking procedure performed in #901 to also
compute the "filtered" ranks. This gives the rest of the metrics in Table 2 of
the ComplEx paper (http://jmlr.org/proceedings/papers/v48/trouillon16.pdf).

As a reminder from #901, the knowledge graph link prediction metrics for a test
edge `E = (s, r, o)` connecting nodes `s` and `o` are calculated by ranking the
prediction for that edge against all modified-source `E' = (n, r, o)` and
modified-object `E'' = (s, r, n)` edges (for all nodes `n` in the graph). The
"raw" ranks are just the rank of `E` against the `E'` and against the `E''`.

The "filtered" ranks exclude the modified edges `E'` and `E''` that are known,
i.e. are in the train, validation or test sets. For instance, if `E = (A, x, B)`
has score 1, but the modified edges `(A, x, C)` and `(A, x, D)` have scores 1.3
and 1.5 respectively, `E` has raw modified-object rank 3. If `(A, x, D)` is in
the train set (or validation or test) but `(A, x, C)` is not, it is excluded
from the filtered ranking, and so `E` has filtered modified-object rank 2.

This has been a struggle to implement correctly, because it has been difficult
to correctly use the right node nodes in the right place of the ranking
procedure. For modified-object ranking, with `E` and `E''` as above, calculating
the score of `E` in the column of scores of every modified-object edge `E''`
needs to use `o`, but calculating the known edges similar to `E` needs to use
`(s, r, _)`, not `(o, r, _)` (the latter is meaningless). (And similarly for
modified-subject ranking.) It sounds obvious when written out like this, but
it's somewhat difficult to keep track of which entity needs to go where in
practice. (@kieranricardo had this key insight.)

The implementation works by: start with the raw `greater` matrix, where each
column represents the test edges `E` with a row for every node in the graph
(i.e. row `n` represents swapping node `n` into the subject or object) and the
elements of a column are `True` if the score of that modified edge is greater
than the score of `E`. For each edge/column, compute the indices of the similar
known edges and set those indices to false, leaving only unknown edges with
scores greater than `E`.

See: #1060