Fix AttributeError, not warning, from DeepGraphInfomax(link_model) (#1149)
The warning message wasn't tested and so had a typo in the code
(`self.base_model` instead of just `base_model`) that triggers an
`AttributeError` exception:

```
AttributeError: 'DeepGraphInfomax' object has no attribute 'base_model'
```

This also rephrases the warning message to focus on what a user will be thinking
about (node vs. link) instead of the implementation detail of the `multiplicity`
property on the models.