Allow flexible walker for unsupervised sampler using new base class RandomWalk (#1187)
This introduces a new base class RandomWalk which is a more specific type 
compared to the existing GraphWalk class. A RandomWalk class is defined as 
having a run method which takes a list of node IDs and returns a list of 
random walks (list of list of node IDs). Currently these classes belong to 
this type:

* UniformRandomWalk
* BiasedRandomWalk
* UniformRandomMetaPathWalk

It's useful to define this type, since these are all suitable for use with 
UnsupervisedSampler, and so we introduce a new walker parameter to 
UnsupervisedSampler to allow for some flexibility instead of being 
hardcoded to use UniformRandomWalk.

See #919

Follow-up #1199