Gather output nodes in full batch models, not the layers (#1204)
The core layer in a full batch method (like `GraphConvolution` inside `GCN`)
natively computes information about a whole graph, but this doesn't directly
work with machine learning (and, in particular, how it works with `Keras`). In
particular, there's usually only a subset of labelled nodes, so training needs
to only compute candidate embeddings for those nodes to compare against the
ground-truth target array.

We currently handle this in full batch methods by having each layer know whether
it is the `final_layer` in the model, and do `tf.gather` using a tensor of
indices of relevant nodes. This means that the individual full-batch layers have
to:

- always accept the tensor of indices, even if they won't use it (and,
  potentially, even if they are being used as part of a model that doesn't ever
  use it)
- do the appropriate `tf.gather` invocation to select the relevant nodes

This patch adjusts this approach to make the output filtering a model level
concern: layers always compute all the information, and models that need to can
do a `tf.gather` call (via the new `layer.misc.GatherIndices`) to select out the
relevant output elements. This has a few benefits:

- it seems conceptually more "correct" to me, because it's the training/using of
  a overall model that needs to filter out elements, _not_ the operation of an
  individual layer
- it reduces the amount of code required noticeably, and will reduce if further
  when we remove the left-over-`final_layer` detection (that is designed to help
  us migrate).
- it makes #1201 easier (in #1205): graph classification doesn't have any output
  indices to feed into the graph convolution layers, and so this patch makes
  `GraphConvolution` closer to the behaviour required for that
- it may also ease implementing models that incorporate information from
  multiple convolution layers, because it's easy to make sure all the
  convolution layers compute embeddings for all nodes

See: #1201