Rename GraphGenerator to PaddedGraphGenerator (#1211)
The `GraphGenerator` name is accurate, because the type generates a sequence of
graphs. However, it is very general, despite it currently performing one
specific style of generating graphs (plopping the features and adjacency
matrices into batched tensors, with padding to make sure the sizes are ok).

Per our off-github discussion, we decided that it is more future-proof to start
with a more specific name than `GraphGenerator`, in case we end up with other
approaches for generating sequences of graphs for training graph classification
algorithms (off the top of my head, possibilities are using ragged tensors, or
some sort of Cluster-GCN-like algorithm that yields equal-sized subgraphs). This
means that our current functionality takes up less "API space", and gives us
more room to manoeuvre.

If we decide that this is the only graph generator type we'll have, we can
always rename (without breaking existing code) in future.

See: #1198