Move CorruptedGenerator and support code to its own file (#1247)
This is now no longer just a full batch generator, so it makes sense to be
separate. It is also a general method that will potentially require more support
code in future. This PR makes no changes to the code, it just moves it verbatim
and updates imports as required.

This is in service of #1176 by making the change required for the generalisation
(#1243) smaller and more clearer: that PR no longer has to do "move code" and
"change code", it can just do the "change code" piece, thus making the
differences clearer.