Generalise Deep Graph Infomax corruption to work with any generator (#1243)
This generalises `DeepGraphInfomax` (in particular, `CorruptedGenerator`) to
theoretically work with any model. It does this by providing a
`corrupt_index_groups` argument, which can be used to control which inputs get
shuffled. It's a list of lists, in the form of a sequence of "groups", where
each group is a sequence of input tensor indices that get concatenated and
shuffled together.

This is a strict generalisation of full-batch shuffling (just shuffle the first
tensor: `[[0]]`) and GraphSAGE shuffling (shuffle all tensors together `[[0, 1,
..., N - 1]]` where `N` is the number of input tensors). HinSAGE fits into the
same framework, by having a group for each type of node (#1173).

For convenience and compatibility, the `corrupt_index_groups` argument can be
set by default from the base generator. The `Generator` class now has an
optional `default_corrupt_input_index_groups` method, which returns `None` by
default, but can be overridden to return a list of appropriate groups for
shuffling. This PR preserves the current `DeepGraphInfomax` functionality by
overriding this for `FullBatchNodeGenerator`, `GraphSAGENodeGenerator` and
`DirectedGraphSAGENodeGenerator`.

The `DeepGraphInfomax` class now has to take in the relevant
`CorruptedGenerator` in order to work out what input tensors to duplicate. The
existing `DeepGraphInfomax(model)` form is still supported for the current set
of classes, but is deprecated.

This doesn't do the fully general form of having a "corrupter" class or function
as in
<https://github.com/stellargraph/stellargraph/pull/1193#issuecomment-607628370>. This
is to avoid creating an abstraction that isn't an appropriate abstraction,
because all of the current algorithms' corruptions are this style of shuffling,
and so we don't have a reference for any other options. If/when we find other
styles of corruption, we can generalise this fairly easily without breaking
backwards compatibility (e.g. add a new argument or check the type of the
existing one).

This also removes the checking and passing of `multiplicity`: it's now up to the
users to pass in an appropriate generator/model for the task they're doing
(_theoretically_ someone could do unsupervised link embedding training or
something, with a link model).

See: #1176