Allow easy unsupervised HinSAGE training with Deep Graph Infomax (#1254)
This uses the generalised Deep Graph Infomax corruption support from #1243 to
have HinSAGE specify a default way to corrupt. This means that HinSAGE can be
easily trained unsupervised via Deep Graph Infomax, by just passing in the
generator and model.

This is currently demonstrated running on Cora, and gives similar results to
GraphSAGE as expected. Creating a real demo significantly more effort, and so is
future work: #1285.

See: #1173