Optimise RelationalFullBatchNodeGenerator construction (18× speedup, 560× smaller) (#1274)
I benchmarked memory usage using pympler, subtracting out the size of the graph
to show the incremental memory use over the raw `StellarGraph` data:

```python
from pympler.asizeof import asizeof
asizeof(sg.mapper.RelationalFullBatchNodeGenerator(G)) - asizeof(G)
```

| code                                          | time (s) | speedup     | memory (MiB) | reduction |
|-----------------------------------------------|----------|-------------|--------------|-----------|
| original (`develop`)                          | 1.64     | 1.0× (1.0×) | 270          | 0%        |
| fetch node features once                      | 1.25     | 1.3× (1.3×) | 270          | 0%        |
| edge-type filtering in `to_adjacency_matrix`  | 0.35     | 4.7× (3.6×) | 270          | 0%        |
| use `_get_index_for_nodes`, not manual `dict` | 0.26     | 6.3× (1.3×) | 269          | -0.4%     |
| don't manually index `node_features`          | 0.09     | 18× (2.9×)  | 0.483        | -99.8%    |

("Speedup" column shows the total speedup vs. `develop`, as well as the
incremental speedup vs. the previous line (in parens). "Reduction" column shows
the total reduction in memory vs `develop`.)

Using `to_adjacency_matrix` means that the directedness of the graph now
matters, so some tests have to be updated to use use `StellarDiGraph` instead of
an undirected `StellarGraph`.

Two of these points require updates to `StellarGraph`:

- edge-type filtering in `to_adjacency_matrix`: expands that method to support
  only including edges of a single type
- don't manually index `node_features`: expands that method to support getting
  all features for a single type, which doesn't have to do any copying (this
  doesn't change behaviour, because the node features were in the right order to
  start with: doing `graph.node_features(graph.nodes())` ends up indexing the
  underlying array with `[0, 1, 2, ... , num_nodes - 1]`, which requires
  allocating new memory and copying, without actually changing the order or
  shape of the data)

The massive memory reduction for the last point is because the feature matrix is
now shared with the underlying `StellarGraph` object and so isn't duplicated in
memory. The final memory usage is thus dominated by the sparse adjacency
matrices for each type, and so should be proportional to `O(number of edge
types + number of edges)` (the first term models the constant cost of each
sparse matrix). Previously, it included a `O(number of nodes * feature size)`
term, which was very large for the AIFB dataset (`8285 nodes * 8285 features per
node * 4 bytes per feature = 262MiB`).

Unfortunately, this still isn't enough to allow the notebook to run on CI, as it
still hits out-of-memory errors (#819):
https://buildkite.com/stellar/stellargraph-public/builds/3220

Some of these optimisations can potentially be applied to other locations in the
library too (#1276, #1277).

See: #1272