Support edges as a DataFrame including the type as a column (#1284)
This simplifies the construction of a `StellarGraph` with multiple edge types:
one can now pass a single DataFrame with a "type column", instead a dictionary
of separate DataFrames for each type.

Many datasets we've worked with come in this format, with all edges in a single
file, with the labels either as a column within that file or as a separate
file. Previously, one would have to separate the edges into a dictionary of
DataFrames (for instance, using a `groupby`) to construct the `StellarGraph`,
but now one can just pass in the DataFrame as is.

This new construction form is activated by passing the `edge_type_column`
parameter to `StellarGraph(...)`. If this argument is passed, the `edges`
parameter must be a single DataFrame, not a dictionary.

This makes constructing graphs with many edge types faster, too. For instance,
the FB15k dataset has almost 600k edges across 1345 types. The time to load it
goes from `5.36 s ± 104 ms` to `2.04 s ± 13.3 ms` (2.6× speedup). Measured by:

```python
import stellargraph as sg
%timeit -n 3 -r 5 sg.datasets.FB15k().load()
```

I think we haven't got quite the right approach for this data conversion yet, as
making changes like this is always a bit clunky.

Follow-up issues:

- #1300 Support `StellarGraph(..., node_type_column=...)` in addition to
  `edge_type_column` for convenient heterogeneity
- #1301 Update basics/ demos to use `StellarGraph(..., edge_type_column=...)` as
  appropriate
- #1302 Use `StellarGraph(..., edge_type_column =...)` in
  `StellarGraph.subgraph`

See: #1183