Install specific version of SG on Colab, in notebooks (#1314)
This uses the versioning logic we've already got for the notebook version checks
(#1242) to install a specific version of StellarGraph on Google Colab. This has
a few benefits:

- it ensures that we can install exactly the version the notebook is designed
  for, even if there's been releases since the notebook was first loaded into
  Colab
- it allows Colab to work with prereleases and release candidates, which pip
  won't install by default
  (https://pip.pypa.io/en/stable/reference/pip_install/#pre-release-versions)

To validate that this is correct syntax, I ran `%pip install
stellargraph[demos]==0.4.0b0` in a jupyter notebook (using a random old
pre-release version we've published). It worked to install that version.

See: #1313