Infer node types for homogeneous-node graphs with 'unique_node_type' (#1375)
This adds a `unique_node_type` method to `StellarGraph`. The API of this is
slightly peculiar, as it either:

- returns the single node type in a graph with only one (a "homogeneous-node
  graph")
- throws a `ValueError` if it doesn't have exactly one node type, optionally
  using a custom error message

This API is what I think is the most reliable one, as it's not possible to
accidentally forget to detect and fail on the case when there's not one node
type. (An alternative API would be to return `None` in the failure case. This
would make sense in some languages, but Python's non-static type system (and
also the flexibility for what StellarGraph allows for node types) means that a
caller could forget to detect this case, and so continue when they should really
stop.)

This PR then uses this new functionality to:

- make `homogeneous_graph.node_features()` work
- make `HinSAGENodeGenerator(homogeneous_graph, batch_size=5, num_samples=[10,
  5])` work, meaning it's now easier to use (#1023) and is a drop-in replacement
  for `GraphSAGE`
- simplify some error handling code in other generators, by calling
  `unique_node_type` instead of manual `if len(...) > 1` code

See: #1343