Simplify "ready for machine learning" out of documentation (#1376)
The "ready for machine learning" phrase is referring to the old "attribute spec"
system for mutating an (old) `StellarGraph` object to insert node features. This
system no longer exists (it was removed in 759b01a), and so this phrase is no
longer relevant. (See #1334 for more specifics.)

We can remove this phrase and simplify our docs in the process for the current
StellarGraph class, where a `StellarGraph` is always ready for machine learning.

The only thing required for some algorithms is features for the nodes. That is,
feature dimension > 0 for each node type. I've phrased this requirement as "with
node features" for brevity, but potentially it would be better to say something
like "with non-empty node features" or "with node features with dimension > 0"
or something along those lines?

See: #1334