Set the right permissions in stellargraph docker image (#1407)
A new release of the `python:3.6-slim` docker image was released
(https://github.com/docker-library/python/commit/bbb44b3ab29), which included an
update to pip 20.1. Per https://pip.pypa.io/en/stable/news:

> Building of local directories is now done in place, instead of a temporary
> location containing a copy of the directory
> tree. (https://github.com/pypa/pip/issues/7555)

This causes our docker build to break, because the `/build` directory was being
created by `WORKDIR`, which means it is owned by `root` (despite being run under
the `stellar` user at the time). This is a problem because we're running `pip
install` on a local directory under the `stellar` user, so pip 20.1 cannot
create directories in that local directory (that is, in `/build`).

This PR pre-creates the `/build` directory and makes sure it's owned by the
`stellar` user, in addition to making sure all the files are copied in under the
same user.

See: #1406