Use commonmark library for validating notebook text, not regexes (#1393)
This is a building block for #1292, since part of that will require validating
the contents of link elements (see: #1394), which will be much more reliable
with a real parser.

I chose the commonmark library https://github.com/readthedocs/commonmark.py
because it (a) provides easy access to an AST node type, and (b) that AST
records the source positions of block elements (in the form of the `sourcepos`
property, which is a list like `[start, end]`, where `start` and `end` are lists
like `[line_number, column_number]` (one-based)). The source positions make it
easy to report nice error messages with lines from the cells, without having to
reconstruct/guess at which part of the code was broken.

This does increase the amount of code. Most of that is in support code, while
the `@checker`s themselves and even error message construction get (arguably)
simpler. The code can now reason in terms of actual markdown entities, instead
of having to do regex parsing/guessing.

This also catches one instance of a heading that the regexes did not, in the
PPNP demo.

Validation: I added some broken-on-purpose notebooks and built them in:
https://buildkite.com/stellar/stellargraph-public/builds/3541

See: #1292