Switch Binder and Colab links in notebooks to markdown block quote (#1398)
The raw HTML table was not converting to reStructuredText through `nbsphinx`
(which, I think, uses `jupyter nbconvert --to=rst`, which uses pandoc). The
output lost the links + images (see more details about `pandoc` in #1398).

This PR updates the links to be normal markdown links + images, which ensures
they do get converted correctly. To distinguish these links from normal body
text, these are formatted as a block-quote (`> ...`). This gives acceptable (not
great) formatting on each of the major platforms one might be using to look at
this (see images in #1398):

- Jupyter locally
- Github:
  https://github.com/stellargraph/stellargraph/blob/2972637ec3b75c7ab7bc6a7a96697489b03f62fd/demos/basics/loading-networkx.ipynb
- nbviewer:
  https://nbviewer.jupyter.org/github/stellargraph/stellargraph/blob/2972637ec3b75c7ab7bc6a7a96697489b03f62fd/demos/basics/loading-networkx.ipynb
- Read the Docs:
  https://stellargraph--1398.org.readthedocs.build/en/1398/demos/basics/loading-networkx.html


Note that both Github and nbviewer have CSS that forces images to be formatted
as block elements, and so they're positioned separately (instead of inline
elements as is the default). This is unfortunate, but per #1391, we're moving
towards Read the Docs being the main online display of our demos.

We can potentially improve the Read the Docs formatting with custom CSS or some
post-processing (similar to #1401).

See: #1293