Fix from_networkx for an empty graph with DataFrame of features (#1379)
This was because it had to find the unique node type that take their features
from the `features` DataFrame. However, without nodes and thus without any node
types, this is impossible.

The approach here is a bit fuzzy and should potentially be reconsidered. For
instance, maybe we should guarantee that there's always at least one type of
nodes and of edges (the default node or edge type) in any `StellarGraph`.

See: #1339