DGI FullBatch methods fix (#1415)
This PR fixes the bug that caused DGI to break on fullbatch when not all nodes were specified. The bug was cause by the fact that fullbatch generators pass in:

(features, target_indicies, adj)

And DGI infers the batch shape from the first input, features in this case, causing the wrong batch shape to be inferred for full batch methods. There are two fixes (that I can think of):

- Specify which input to infer the batch shape from for each generator along with num_batch_dims
- Take the minimum batch shape along all inputs

This PR makes the second change as a minimal fix - minimal code and test changes - to get things working quickly but might be too hacky. Thoughts?

See #1349

* fixed bug

* clarify output batch_shape