Add GCNSupervisedGraphClassification(pool_all_layers=...) to share with DGCNN (#1373)
This adds the `pool_all_layers` parameter to `GCNSupervisedGraphClassification`
to include the output of every GCN layer in the pooling aggregation.

Between this and the custom pooling of #1369, we can merge the actual
implementation of DGCNN with that of `GCNSupervisedGraphClassification`. The
`DeepGraphConvolutionalNeuralNetwork`/DGCNN class just needs to set the
parameters for the `GCNSupervisedGraphClassification` superclass construction
appropriately. This reduces how much specialised code, documentation and testing
is required.

This also adds a unit test that does a few little operations to check things are
composed correctly. This probably doesn't need much more tests because it is
such a thin layer around `SortPooling` and `GCNSupervisedGraphClassification`,
both of which have good tests.

In particular, the test for `DeepGraphConvolutionalNeuralNetwork` checks that:

- the class can be constructed
- `GCNSupervisedGraphClassification`'s parameters seem to be set correctly
- the predictions have an appropriate shape

(For consistency/precision, it renames the
`test_gcn_supervised_graph_classification.py` file to
`test_graph_classification.py`, to match `graph_classification.py`.)

See: #1371 and #1297