Add a diff to format_notebooks output on CI, for diagnosis (#1421)
In builds like https://buildkite.com/stellar/stellargraph-public/builds/3569 and
https://buildkite.com/stellar/stellargraph-public/builds/3573 on #1253, we
observed an unexplained formatting failure, where ~every notebook was apparently
formatted incorrectly. The first commit that hit this was
https://github.com/stellargraph/stellargraph/pull/1253/commits/11fd94010a948c0ffff7b6a014dd772f6d05a7b3
which didn't change any notebooks or the `format_notebooks.py` script.

CI doesn't have enough state to be able to work out what's going on. For
instance, we can't get the final formatted notebook to do a
comparison. Including a diff is the simplest way to get more visibility into
this. (Another option would be uploading the before/after for failing notebooks
as Buildkite artifacts, but that's a bit more fiddly, and, a human is likely to
just download them and do a diff manually anyway.)

Example: I manually broke the formatting of a notebook in
https://github.com/stellargraph/stellargraph/pull/1421/commits/6a3216959faf45b046bbbe76186b72706f5a3515
. A CI run on that commit contains output like
https://buildkite.com/stellar/stellargraph-public/builds/3644#036dc2ba-e920-46a3-afd0-6fac24464484/159-183