Rename DeepGraphConvolutionalNeuralNetwork to DeepGraphCNN (#1422)
The paper https://www.cse.wustl.edu/~muhan/papers/AAAI_2018_DGCNN.pdf that
introduces our (currently-named) `DeepGraphConvolutionalNeuralNetwork` calls it
"DGCNN". This abbreviation is nice and short and matches both the paper and the
convention in the rest of the library (e.g. `GCN` instead of
`GraphConvolutionalNetwork`). It also avoids redundancy like "Graph" in a
library entirely designed around graphs, and expanding common terms-of-art like
"Neural Network" and even "CNN"
https://en.wikipedia.org/wiki/Convolutional_neural_network

Unfortunately, the DGCNN acronym has a collision, with "Dynamic Graph CNN"
https://github.com/WangYueFt/dgcnn

This PR renames the class to `DeepGraphCNN` to balance brevity and convention
with specificity.

See: #1370