Allow images within links in notebooks (#1442)
Pull requests #1398 (using pure markdown for colab links) and #1394 (no nested
formatting in notebooks) merged at about the same time, but they conflicted. The
syntax used in #1398 failed the check of #1394, and indeed revealed a missing
piece: it's ok to have a link that contains an image in notebook markdown,
because this passes through the reStructuredText conversion ok.

(This was yet more "merge skew".)

This also has to replace an instance of ***bold-italic*** with plain **bold** in
the GCN-LSTM notebook, because the CI requires that this doesn't appear (as it
doesn't render correctly on Read the Docs; #1394 discusses this in more detail).