Disallow leading-block-quote in cells, allow summary in title cell (#1413)
This works around problems in the Markdown to reStructuredText conversion that
nbsphinx does when rendering notebooks. In particular, spatialaudio/nbsphinx#450
and/or jupyter/nbconvert#1245. These issues mean that if a cell starts with a
block quote, it will merge with a reStructuredText directive at the end of the
previous cell, because an appropriate separating comment is not inserted.

Pull request #1398 gives an example of this; the `loading-networkx.ipynb`
notebook's rST from `nbconvert` ends up looking like:

```rst
   Run the master version of this notebook: |Open In Binder| |Open In
   Colab|

.. |Open In Binder| image:: https://mybinder.org/badge_logo.svg
   :target: https://mybinder.org/v2/gh/stellargraph/stellargraph/master?urlpath=lab/tree/demos/basics/loading-networkx.ipynb
.. |Open In Colab| image:: https://colab.research.google.com/assets/colab-badge.svg
   :target: https://colab.research.google.com/github/stellargraph/stellargraph/blob/master/demos/basics/loading-networkx.ipynb

   This demo explains how to load data from NetworkX into a form that
   can be used by the StellarGraph library. `See all other
   demos <../README.md>`__.
```

The "This demo ..." text gets included as part of the `|Open In Colab|`
substitution definition. This notebook thus isn't displayed correct, and
potentially has warnings/errors when rendering. If this was all in a single
cell, the rST would look like:

```rst
   Run the master version of this notebook: |Open In Binder| |Open In
   Colab|

.. |Open In Binder| image:: https://mybinder.org/badge_logo.svg
   :target: https://mybinder.org/v2/gh/stellargraph/stellargraph/master?urlpath=lab/tree/demos/basics/loading-networkx.ipynb
.. |Open In Colab| image:: https://colab.research.google.com/assets/colab-badge.svg
   :target: https://colab.research.google.com/github/stellargraph/stellargraph/blob/master/demos/basics/loading-networkx.ipynb

..

   This demo explains how to load data from NetworkX into a form that
   can be used by the StellarGraph library. `See all other
   demos <../README.md>`__.
```

Where the extra `..` is an empty comment, which is the recommended way to
separate a quote from an earlier construct
(https://docutils.sourceforge.io/docs/ref/rst/restructuredtext.html#block-quotes).

This pull request fixes this in two ways:

- flagging any markdown cells that have a block quote at the start, since they
  may hit these problems (except for the cloud-runner ones, which are entirely a
  quote now #1398)
- allowing the first cell to also include a summary after title, where the
  summary is denoted by a `> ...` quote element

Examples of the failures (on the notebooks that were fixed here):
https://buildkite.com/stellar/stellargraph-public/builds/3613

Comparison: (see #1413 for images)

- Before: The notebook summary is missing.
  https://stellargraph--1398.org.readthedocs.build/en/1398/demos/basics/loading-pandas.html
- After: The notebook summary is included.
  https://stellargraph--1413.org.readthedocs.build/en/1413/demos/basics/loading-pandas.html

Also see:
https://github.com/stellargraph/stellargraph/pull/1398#issuecomment-621606354

See: #1293