Optimise weighted BiasedRandomWalk (4× speedup) (#1444)
The switch to using ilocs for edges and thus in the random walkers (#1267
991db681) resulted in some dramatic slowdowns in the `BiasedRandomWalk`
code. This is mainly due to requiring more ID<->iloc conversions for some
methods. Some of that overhead is unavoidable. This PR removes some of the
avoidable overhead.

This increases the size of the `BiasedRandomWalk` benchmark, so that they're
slightly more realistic without taking ridiculously long. In particular, it's
rare that someone would do a random walk starting at a single node.

Given that, it then optimises the `BiasedRandomWalk` code, focusing on the
weighted form, because of the large slowdown observed in the change to use ilocs
(from f2a96aab to 991db681).

The following table reports the median times for the `` and `` benchmarks (as
run by `pytest -k 'benchmark and biased'`), along with the speedups
(vs. `develop`), and incremental speedups (vs. the previous row).

| change                                          | unweighted (ms) | speedup (incr.) | weighted (ms) | speedup (incr.) |
|-------------------------------------------------|-----------------|-----------------|---------------|-----------------|
| f2a96aab (pre-iloc), new benchmark              | 8.58            | 2.0× (N/A)      | 86.7          | 1.6×            |
| `develop`, new benchmark                        | 17.5            | 1.0× (N/A)      | 143           | 1.0× (N/A)      |
| numpy-ised weight validation                    | 17.5            | 1.0× (1.0×)     | 75.0          | 1.9× (1.9×)     |
| do `length` `neighbors` calls, not `length + 1` | 15.4            | 1.1× (1.1×)     | 71.1          | 2.0× (1.05×)    |
| move `def transition_probability`               | 15.4            | 1.1× (1.0×)     | 70.9          | 2.0× (1.0×)     |
| use `include_edge_weights`, not `_edge_weights` | 15.4            | 1.1× (1.0×)     | 36.8          | 3.9× (1.9×)     |

That is, weighted random walks with the new form sees a speedup of almost 4×
over the current state of `develop`, and 2.4× over the state before the iloc
change. This recovers only a slight amount of the performance regression for
unweighted walks.

Here's the numbers on a "real world" dataset (Cora), comparing this PR, with the
state when ilocs were introduced and the state immediately before that. Measured
using:

```python
import stellargraph as sg
G, _ = sg.datasets.Cora().load()
rw = sg.data.BiasedRandomWalk(G, n=3, length=10, p=2, q=4)
%timeit -n 1 -r 3 rw.run(G.nodes())
%timeit -n 1 -r 3 rw.run(G.nodes(), weighted=True)
```

(NB. the weighted walks are equivalent to the unweighted ones, given they're all
the default weight of 1, but it's still using the weighted code path.)

| code                | unweighted (s) | weighted (s)  |
|---------------------|----------------|---------------|
| f2a96aab (pre-iloc) | 1.36 ± 0.0233  | 11.7 ± 0.0724 |
| 991db681 (iloc)     | 2.02 ± 0.0173  | 13 ± 0.427    |
| This PR             | 2 ± 0.101      | 5.64 ± 0.179  |

The speedups/slowdowns are less dramatic here, because there's more real work,
and so the constant overhead of the initial iloc conversion and the weight
validation is less noticeable.

See: #1441